third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-first-letter-first-line-hover.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
#p { color: red }
#p::first-line { color: green }
#p::first-letter { background-color: lime; }
#p:hover { text-transform: lowercase }
</style>
<p id="p">Hover me.</p>
<script>
var green = "rgb(0, 128, 0)";
shouldBe("getComputedStyle(p, '::first-letter').color", "green");

if (eventSender)
    eventSender.mouseMoveTo(p.offsetLeft+1, p.offsetTop+1);

shouldBe("getComputedStyle(p, '::first-letter').color", "green");
</script>
