third_party/WebKit/LayoutTests/fast/css/getComputedStyle/getComputedStyle-first-letter-first-line-hover-expected.txt

PASS getComputedStyle(p, '::first-letter').color is green
FAIL getComputedStyle(p, '::first-letter').color should be rgb(0, 128, 0). Was rgb(255, 0, 0).
PASS successfullyParsed is true

TEST COMPLETE
hover me.
